feat(trace): move to opentracing package
