+++ pkg/trace/module.go
@@ -1,8 +1,14 @@
 package trace
 
 import (
+	"github.com/alexfalkowski/go-service/pkg/trace/opentracing"
 	"go.uber.org/fx"
 )
 
-// Module for fx.
-var Module = fx.Invoke(Register)
+var (
+	// JaegerOpenTracing for fx.
+	JaegerOpenTracing = fx.Invoke(opentracing.RegisterJaeger)
+
+	// Module for fx.
+	Module = fx.Options(JaegerOpenTracing)
+)
